requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.8.5
Flask-JWT-Extended==4.5.3
marshmallow==3.20.1
bcrypt==4.0.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5